build: optional tokio_unstable support
This makes using the tokio_unstable feature optional - the build
configuration defines what is enabled, rather than the other way around.

By default, and for prod builds and users who have not set a RUSTFLAGS,
the tokio_unstable is enabled and this is leveraged to provide the
metrics exposed by the flag. For builds where tokio_unstable is not
enabled, these metrics are not included, without causing a compilation
error.

CI will against the flag-enabled/prod code by virtue of being the
default.
