+++ influxdb_iox/src/process_info.rs
@@ -3,7 +3,8 @@ use std::sync::Arc;
 use iox_time::{SystemProvider, Time, TimeProvider};
 use metric::U64Gauge;
 use once_cell::sync::Lazy;
-use tokio::runtime::Handle;
+
+#[cfg(tokio_unstable)]
 use tokio_metrics_bridge::setup_tokio_metrics;
 
 /// Package version.
@@ -54,7 +55,12 @@ pub fn setup_metric_registry() -> Arc<metric::Registry> {
     registry.register_instrument("jemalloc_metrics", crate::jemalloc::JemallocMetrics::new);
 
     // Register tokio metric for main runtime
-    setup_tokio_metrics(Handle::current().metrics(), "main", Arc::clone(&registry));
+    #[cfg(tokio_unstable)]
+    setup_tokio_metrics(
+        tokio::runtime::Handle::current().metrics(),
+        "main",
+        Arc::clone(&registry),
+    );
 
     registry
 }
